// types/IRegistrarEmpleado.ts
// Generated by https://quicktype.io

export interface IRegistarEmpleado {
    nombre:                       string;
    apellidos:                    string;
    tipoDocumento:                number;
    numeroDocumento:              string;
    fechaNacimiento:              string;
    genero:                       number;
    estadoCivil:                  number;
    nacionalidad:                 null;
    correo:                       null;
    telefonoMovil:                null;
    direccion:                    null;
    distrito:                     string;
    provincia:                    string;
    departamento:                 string;
    contactoEmergenciaNombre:     null;
    contactoEmergenciaParentesco: number;
    contactoEmergenciaTelefono:   null;
    numeroCuentaBancaria:         null;
    bancoNombre:                  null;
    tipoCuenta:                   number;
    cci:                          null;
    ruc:                          null;
    numeroESSalud:                null;
    sistemaPensiones:             number;
    cuspp:                        null;
    nivelEducativo:               number;
    profesionOficio:              null;
    fotoUrl:                      string | null;
    cargoId:                      number;
    salario:                      number;
    tipoContrato:                 number;
    tipoJornada:                  number;
    fechaIngreso:                 string;
    observaciones:                null;
}